Updated about me and moved Portfolio to routes folder
